Basics

Basics

Dynamical Systems Utilities

Definition of vector fields

CoherentStructures.jl is set up for handling two- and three-dimensional dynamical systems only. For such low-dimensional flows it is advantageous (for top performance) to obey the following syntax:

function vectorfield2d(u,p,t)
    du1 = ... # equation for $\dot{x}$
    du2 = ... # equation for $\dot{y}$
    return StaticArrays.SVector{2}(du1,du2)
end

and correspondingly for three-dimensional ODEs:

function vectorfield3d(u,p,t)
    du1 = ... # equation for $\dot{x}$
    du2 = ... # equation for $\dot{y}$
    du3 = ... # equation for $\dot{z}$
    return StaticArrays.SVector{3}(du1,du2,du3)
end

Furthermore, there is a convenience macro to define two-dimensional velocity fields from stream functions.

CoherentStructures.@define_stream — Macro.

@define_stream(name::Symbol, code::Expr)

Define a scalar stream function on $R^2$. The defining code can be a series of definitions in an enclosing begin ... end-block and is treated as a series of symbolic substitutions. Starting from the symbol name, substitutions are performed until the resulting expression only depends on x, y and t.

The symbol name is not brought into the namespace. To access the resulting vector field and variational equation use @velo_from_stream name and @var_velo_from_stream name

This is a convenience macro for the case where you want to use @velo_from_stream and @var_velo_from_stream without typing the code twice. If you only use one, you might as well use @velo_from_stream name code or @var_velo_from_stream directly.

source

CoherentStructures.@velo_from_stream — Macro.

@velofromstream(name::Symbol, [code::Expr])

Get the velocity field corresponding to a stream function on $R^2$. The defining code can be a series of definitions (in an enclosing begin ... end-block and is treated as a series of symbolic substitutions. Starting from the symbol name, substitutions are performed until the resulting expression only depends on x, y and t.

The macro returns an anonymous function f(u,p,t) that returns a SVector{2} holding the vector field at u at time t.

The macro can be called without the code if the stream name has been define beforehand via @define_stream.

Examples

julia> using CoherentStructures

julia> f = @velo_from_stream Ψ_ellipse begin
               Ψ_ellipse = a*x^2 + b*y^2
               a = t
               b = 3
           end
(#3) (generic function with 1 method)

julia> f([1.0,1.0], nothing, 1.0)
2-element StaticArrays.SArray{Tuple{2},Float64,1,2}:
 -6.0
  2.0

julia> using CoherentStructures

julia> @define_stream Ψ_circular begin
           Ψ_circular = f(x) + g(y)

           # naming of function variables
           # does not matter:
           f(a) = a^2
           g(y) = y^2
       end

julia> f2 = @velo_from_stream Ψ_circular
(#5) (generic function with 1 method)

julia> f2([1.0,1.0], nothing, 0.0)
2-element StaticArrays.SArray{Tuple{2},Float64,1,2}:
 -2.0
  2.0

source

In fact, two of the predefined velocity fields, the rotating double gyre rot_double_gyre, and the Bickley jet flow bickleyJet, are generated from these macros.

Another typical use case is when velocities are given as a data set. In this case, one first interpolates the velocity components with interpolateVF to obtain callable interpolation functions, say, UI and VI. The corresponding vector field is then interp_rhs, which the velocity interpolants enter via the parameter argument p; see below for examples.

CoherentStructures.interpolateVF — Function.

interpolateVF(xspan,yspan,tspan,u,v,interpolation_type=ITP.BSpline(ITP.Cubic(ITP.Free())))) -> UI, VI

source

CoherentStructures.interp_rhs — Function.

interp_rhs(u, p, t) -> SA.SVector{2}

Defines a 2D vector field that is readily usable for trajectory integration from vector field interpolants of the x- and y-direction, resp. It assumes that the interpolants are provided as a 2-tuple (UI, VI) via the parameter p. Here, UI and VI are the interpolants for the x- and y-components of the velocity field.

source

Flow maps

CoherentStructures.flow — Function.

flow(odefun,  u0, tspan; tolerance, p, solver) -> Vector{Vector}

Solve the ODE with right hand side given by odefun and initial value u0. p is a parameter passed to odefun. tolerance is passed as both relative and absolute tolerance to the solver, which is determined by solver.

source

CoherentStructures.parallel_flow — Function.

parallel_flow(flow_fun,P) -> Array

Apply the flow_fun to each element in P in parallel, if possible. Returns a 2D array with dimensions ((space dim x no. of time instances) x no. of trajectories), in which each column corresponds to a concatenated trajectory, i.e., represented in delay coordinates.

source

Linearized flow map

CoherentStructures.linearized_flow — Function.

linearized_flow(odefun, x, tspan,δ; ...) -> Vector{Tensor{2,2}}

Calculate derivative of flow map by finite differences if δ != 0. If δ==0, attempts to solve variational equation (odefun is assumed to be the rhs of variational equation in this case). Return time-resolved linearized flow maps.

source

Cauchy-Green and other pullback tensors

CoherentStructures.parallel_tensor — Function.

parallel_tensor(tensor_fun,P) -> Array{SymmetricTensor}

Computes a tensor field via tensor_fun for each element of P, which is an array of vectors. tensor_fun is a function that takes initial conditions as input and returns a symmetric tensor. The final tensor field array has the same size as P.

source

CoherentStructures.CG_tensor — Function.

CG_tensor(odefun, u, tspan, δ; kwargs...) -> SymmetricTensor

Returns the classic right Cauchy–Green strain tensor. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

kwargs...: are passed to linearized_flow

source

CoherentStructures.mean_diff_tensor — Function.

mean_diff_tensor(odefun, u, tspan, δ; kwargs...) -> SymmetricTensor

Returns the averaged diffusion tensor at a point along a set of times. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

kwargs...: are passed to linearized_flow

source

CoherentStructures.av_weighted_CG_tensor — Function.

av_weighted_CG_tensor(odefun, u, tspan, δ; G, kwargs...) -> SymmetricTensor

Returns the transport tensor of a trajectory, aka time-averaged, di ffusivity-structure-weighted version of the classic right Cauchy–Green strain tensor. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

G: (constant) metric tensor

kwargs... are passed through to linearized_flow

source

CoherentStructures.pullback_tensors — Function.

pullback_tensors(odefun, u, tspan, δ; D, kwargs...) -> Tuple(Vector{SymmetricTensor},Vector{SymmetricTensor})

Returns the time-resolved pullback tensors of both the diffusion and the metric tensor along a trajectory. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

D: (constant) diffusion tensor, metric tensor is computed via inversion; defaults to eye(2)

kwargs... are passed through to linearized_flow

source

CoherentStructures.pullback_metric_tensor — Function.

pullback_metric_tensor(odefun, u, tspan, δ; G, kwargs...) -> Vector{SymmetricTensor}

Returns the time-resolved pullback tensors of the metric tensor along a trajectory, aka right Cauchy-Green strain tensor. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

G: (constant) metric tensor

kwargs... are passed through to linearized_flow

source

CoherentStructures.pullback_diffusion_tensor — Function.

pullback_diffusion_tensor(odefun, u, tspan, δ; D, kwargs...) -> Vector{SymmetricTensor}

Returns the time-resolved pullback tensors of the diffusion tensor along a trajectory. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

D: (constant) diffusion tensor

kwargs... are passed through to linearized_flow

source

CoherentStructures.pullback_SDE_diffusion_tensor — Function.

pullback_SDE_diffusion_tensor(odefun, u, tspan, δ; D, kwargs...) -> Vector{SymmetricTensor}

Returns the time-resolved pullback tensors of the diffusion tensor in SDEs. Derivatives are computed with finite differences.

odefun: RHS of the ODE

u: initial value of the ODE

tspan: set of time instances at which to save the trajectory

δ: stencil width for the finite differences

D: (constant) diffusion tensor

kwargs... are passed through to linearized_flow

source

A second-order symmetric two-dimensional tensor field may be diagonalized pointwise ("eigendecomposition") by the following function.

CoherentStructures.tensor_invariants — Function.

tensor_invariants(T) -> λ₁, λ₂, ξ₁, ξ₂, traceT, detT

Returns pointwise invariants of the 2D symmetric tensor field T, i.e., smallest and largest eigenvalues, corresponding eigenvectors, trace and determinant.

Example

T = [Tensors.SymmetricTensor{2,2}(rand(3)) for i in 1:10, j in 1:20]
λ₁, λ₂, ξ₁, ξ₂, traceT, detT = tensor_invariants(T)

All output variables have the same array arrangement as T; e.g., λ₁ is a 10x20 array with scalar entries.

source

Distance computations

To compute distances w.r.t. standard metrics, there exists the excellent Distance.jl package. The Euclidean distance between two points is computed by any of the following lines:

using Distances
x, y = rand(3), rand(3)
evaluate(Euclidean(),x,y)
euclidean(x,y)

Other metrics of potential interest include Haversine(r), the geodesic distance of two points on the sphere with radius r. In CoherentStructures.jl, there are two more types of metrics implemented:

CoherentStructures.PEuclidean — Type.

PEuclidean(L)

Create a Euclidean metric on a rectangular periodic domain. Periods per dimension are contained in the vector L. For dimensions without periodicity put Inf in the respective component.

Example

julia> x, y, L = [0.0, 0.0], [0.7, 0.0], [0.5, Inf]
([0.0, 0.0], [0.7, 0.0], [0.5, Inf])

julia> Distances.evaluate(PEuclidean(L),x,y)
0.19999999999999996

source

CoherentStructures.STmetric — Type.

STmetric(Smetric, dim, p)

Creates a spatiotemporal, averaged in time metric.

Properties

Smetric is a metric as defined in the Distances package, e.g., Euclidean, PEuclidean, or Haversine;

dim corresponds to the spatial dimension;

p corresponds to the kind of average applied to the vector of spatial distances:

p = Inf: maximum

p = 2: mean squared average

p = 1: arithmetic mean

p = -1: harmonic mean (does not yield a metric!)

p = -Inf: minimum (does not yield a metric!)

Example

julia> x, y = rand(10), rand(10)

julia> Distances.evaluate(STmetric(Distances.Euclidean(),2,1),x,y)

source

That is, the distance on a periodic torus/cylinder, and a spatiotemporal metric that interprets vectors as concatenated trajectories, applies the spatial metric to each time instance and reduces the vector of spatial distances by computing its $l_p$-norm. Notably, p may be any "real" number, including Inf and -Inf for the maximum- and "minimum"-norm. The spatiotemporal metric is a mathematical metric only for $p\geq 1$, in which case it smoothly operates with efficient sparsification methods like BallTree and inrange as implemented in the NearestNeighbors.jl package.