Geodesic vortices

Geodesic elliptic material vortices

The following functions implement an LCS methodology developed in the following papers:

Haller & Beron-Vera, 2012

Haller & Beron-Vera, 2013

Karrasch, Huhn, and Haller, 2015

The present code is structurally inspired–albeit partially significantly improved–by Alireza Hadjighasem's MATLAB implementation, which was written in the context of the SIAM Review paper. Depending on the indefinite metric tensor field used, the functions below yield the following types of coherent structures:

black-hole/Lagrangian coherent vortices (Haller & Beron-Vera, 2012)

elliptic objective Eulerian coherent structures (OECSs) (Serra & Haller, 2016)

material diffusive transport barriers (Haller, Karrasch, and Kogelbauer, 2018)

The general procedure is the following. Assume $T$ is the symmetric tensor field of interest, say, (i) the Cauchy-Green strain tensor field $C$, (ii) the rate-of-strain tensor field $S$, or (iii) the averaged diffusion-weighted Cauchy-Green tensor field $\bar{C}_D$; cf. the references above. Denote by $0<\lambda_1\leq\lambda_2$ the eigenvalue and by $\xi_1$ and $\xi_2$ the corresponding eigenvector fields of $T$. Then the direction fields of interest are given by

\[\eta_{\lambda}^{\pm} := \sqrt{\frac{\lambda_2 - \lambda}{\lambda_2-\lambda_1}}\xi_1 \pm \sqrt{\frac{\lambda - \lambda_1}{\lambda_2-\lambda_1}}\xi_2.\]

Tensor singularities are defined as points at which $\lambda_2=\lambda_1$, i.e., at which the two characteristic directions $\xi_1$ and $\xi_2$ are not well-defined. Then, the algorithm put forward in Karrasch et al., 2015 consists of the following steps:

locate singularities of the tensor field $T$;

determine the type of the singularities (only non-degenerate types like wedges and trisectors are detected);

look for wedge pairs that are reasonably isolated from other singularities;

place an eastwards oriented Poincaré section at the pair center;

for each point on the discretized Poincaré section, scan through the parameter space such that the corresponding η-orbit closes at that point;

for each Poincaré section, take the outermost closed orbit as the coherent vortex barrier (if there exist any).

Function documentation

The fully automated meta-function is the following:

CoherentStructures.ellipticLCS — Function.

ellipticLCS(T,xspan,yspan,p)

Computes elliptic LCSs as null-geodesics of the Lorentzian metric tensor field given by shifted versions of T on the 2D computational grid spanned by xspan and yspan. p is a tuple of the following parameters (in that order):

radius: radius in tensor singularity type detection,

MaxWdgeDist: maximal distance to nearest wedge-type singularity,

MinWedgeDist: minimal distance to nearest wedge-type singularity,

Min2ndDist: minimal distance to second-nearest wedge-type singularity,

p_length: length of Poincaré section,

n_seeds: number of seeding points along the Poincaré section,

Returns a list of tuples, each tuple containing

the parameter value λ in the η-formula,

the sign used in the η-formula,

the outermost closed orbit for the corresponding λ and sign.

source

Essentially, it calls sequentially the following functions.

CoherentStructures.singularity_location_detection — Function.

singularity_location_detection(T,xspan,yspan)

Detects tensor singularities of the tensor field T, given as a matrix of SymmetricTensor{2,2}. xspan and yspan correspond to the uniform grid vectors over which T is given. Returns a list of static 2-vectors.

source

CoherentStructures.singularity_type_detection — Function.

singularity_type_detection(singularity,T,radius,xspan,yspan)

Determines the singularity type of the singularity candidate singularity by querying the tensor eigenvector field of T in a circle of radius radius around the singularity. xspan and yspan correspond to the computational grid. Returns 1 for a trisector, -1 for a wedge, and 0 otherwise.

source

CoherentStructures.detect_elliptic_region — Function.

detect_elliptic_region(singularities,singularityTypes,MaxWedgeDist,MinWedgeDist,Min2ndDist)

Determines candidate regions for closed tensor line orbits.

singularities: list of all singularities

singularityTypes: list of corresponding singularity types

MaxWedgeDist: maximum distance to closest wedge

MinWedgeDist: minimal distance to closest wedge

Min2ndDist: minimal distance to second closest wedge

Returns a list of vortex centers.

source

CoherentStructures.set_Poincaré_section — Function.

set_Poincaré_section(vc,p_length,n_seeds,xspan,yspan)

Generates a horizontal Poincaré section, centered at the vortex center vc of length p_length consisting of n_seeds starting at 0.2*p_length eastwards. All points are guaranteed to lie in the computational domain given by xspan and yspan.

source

CoherentStructures.compute_outermost_closed_orbit — Function.

compute_outermost_closed_orbit(pSection, T, xspan, yspan; pmin = .7, pmax = 1.3)

Compute the outermost closed orbit for a given Poincaré section pSection, tensor field T, where the total computational domain is spanned by xspan and yspan. Keyword arguments pmin and pmax correspond to the range of shift parameters in which closed orbits are sought.

source