Graph Laplacian-based methods

Graph Laplacian/diffusion maps-based LCS methods

The LCS approaches implemented and described in this section are strongly influenced by ideas developed in the spectral clustering/diffusion maps communities. The general references include:

Shi & Malik, Normalized cuts and image segmentation, 2000

Coifman & Lafon, Diffusion maps, 2006

Marshall & Hirn, Time coupled diffusion maps, 2018

In the LCS context, these ideas have been adopted in the following works:

somewhat related Froyland & Padberg-Gehle, 2015

Hadjighasem et al., 2016

Banisch & Koltai, 2017

Rypina et al., 2017/Padberg-Gehle & Schneide, 2018

For demonstrations on example cases, please consult the page on Working with trajectories.

Function documentation

Sparsification methods

Three commonly used sparsification methods are implemented for use with various graph Laplacian methods.

CoherentStructures.KNN — Type.

KNN(k) <: SparsificationMethod

Defines the KNN (k-nearest neighbors) sparsification method. In this approach, first k nearest neighbors are sought. In the final graph Laplacian, only those particle pairs are included which are contained in some k-Neighborhood.

source

CoherentStructures.MutualKNN — Type.

MutualKNN(k) <: SparsificationMethod

Defines the mutual KNN (k-nearest neighbors) sparsification method. In this approach, first k nearest neighbors are sought. In the final graph Laplacian, only those particle pairs are included which are mutually contained in each others k-Neighborhood.

source

CoherentStructures.Neighborhood — Type.

Neighborhood(ε) <: SparsificationMethod

Defines the ε-Neighborhood sparsification method. In the final graph Laplacian, only those particle pairs are included which have distance less than ε.

source

Other sparsification methods can be implemented by defining a SparsificationMethod type and a corresponding spdist method.

Diffusion-maps-type/adjancency-matrix-based graph Laplacian methods

Since the Euclidean heat kernel is ubiquitous in diffusion maps-based computations, we provide it for convenience.

CoherentStructures.gaussian — Function.

gaussian(σ)

Returns the Euclidean heat kernel as a callable function

\[x \mapsto \exp\left(-\frac{x^2}{4\sigma}\right)\]

Example

julia> kernel = gaussian(2.0);

julia> kernel(0.)
1.0

source

CoherentStructures.gaussiancutoff — Function.

gaussiancutoff(σ, θ)

Computes the positive value at which gaussian equals θ, i.e.,

\[\sqrt{-4\sigma\log(\theta)}\]

source

To compute a sparse distance matrix (or adjacency matrix, depending on the sparsification method), use spdist.

CoherentStructures.spdist — Function.

spdist(data, sp_method, metric=Euclidean()) -> SparseMatrixCSC

Return a sparse distance matrix as determined by the sparsification method sp_method and metric.

source

The main high-level functions for the above-listed methods are the following.

CoherentStructures.sparse_diff_op_family — Function.

sparse_diff_op_family(data, sp_method, kernel, op_reduce; α, metric, verbose)

Return a list of sparse diffusion/Markov matrices P.

Arguments

data: a list of trajectories, each a list of states of type SVector;

sp_method: a sparsification method;

kernel=gaussian(): diffusion kernel, e.g., gaussian;

op_reduce=P -> prod(LinearMaps.LinearMap, reverse(P)): time-reduction of diffusion operators, e.g. Statistics.mean (space-time diffusion maps), P -> row_normalize!(max.(sum(P), 1)) (network-based coherence) or the default P -> prod(LinearMaps.LinearMap, reverse(P)) (time coupled diffusion maps)

Keyword arguments

α=1: exponent in diffusion-map normalization;

metric=Euclidean(): distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$;

verbose=false: whether to print intermediate progress reports.

source

CoherentStructures.sparse_diff_op — Function.

sparse_diff_op(data, sp_method, kernel; α=1, metric=Euclidean()) -> SparseMatrixCSC

Return a sparse diffusion/Markov matrix P.

Arguments

data: a list of trajectories, each a list of states of type SVector, or a list of states of type SVector;

sp_method: a sparsification method;

kernel: diffusion kernel, e.g., gaussian;

Keyword arguments

α: exponent in diffusion-map normalization;

metric: distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$.

source

Normalization functions

In the diffusion maps framework, there are two commonly used normalization steps:

kernel-density estimate normalization (kde_normalize!), and

row normalization (row_normalize!), to obtain a diffusion/Markov operator (w.r.t. right- and left-action, respectively).

CoherentStructures.kde_normalize! — Function.

kde_normalize!(A, α=1)

Normalize rows and columns of A in-place with the respective row-sum to the α-th power; i.e., return $a_{ij}:=a_{ij}/q_i^{\alpha}/q_j^{\alpha}$, where $q_k = \sum_{\ell} a_{k\ell}$. Default for α is 1.

source

CoherentStructures.row_normalize! — Function.

row_normalize!(A)

Normalize rows of A in-place with the respective row-sum; i.e., return $a_{ij}:=a_{ij}/q_i$.

source

Diffusion-coordinate-like functions

CoherentStructures.diffusion_coordinates — Function.

diffusion_coordinates(P, n_coords; maxiter=3000) -> (Σ::Vector, Ψ::Matrix)

Compute the (time-coupled) diffusion coordinates Ψ and the coordinate weights Σ for a diffusion operator P. n_coords determines the number of diffusion coordinates to be computed, maxiter is passed to Arpack.eigs.

source

CoherentStructures.diffusion_distance — Function.

diffusion_distance(diff_coord) -> SymmetricMatrix

Returns the distance matrix of pairs of points whose diffusion distances correspond to the diffusion coordinates given by diff_coord.

source