Graph Laplacian-based methods

Graph Laplacian/diffusion maps-based LCS methods

Cite a couple of important papers:

Shi & Malik, Normalized cuts and image segmentation, 2000

Coifman & Lafon, Diffusion maps, 2006

Marshall & Hirn, Time coupled diffusion maps, 2018

In the LCS context, we have

somewhat related Froyland & Padberg-Gehle, 2015

Hadjighasem et al., 2016

Banisch & Koltai, 2017

Rypina et al., 2017/Padberg-Gehle & Schneide, 2018

De Diego et al., 2018

Function documentation

Sparsification methods

Two commonly used sparsification methods are implemented for use with various graph Laplacian methods, see below.

CoherentStructures.KNN — Type.

KNN(k)

Defines the KNN (k-nearest neighbors) sparsification method. In this approach, first k nearest neighbors are sought. In the final graph Laplacian, only those particle pairs are included which are contained in some k-neighborhood.

source

CoherentStructures.mutualKNN — Type.

mutualKNN(k)

Defines the mutual KNN (k-nearest neighbors) sparsification method. In this approach, first k nearest neighbors are sought. In the final graph Laplacian, only those particle pairs are included which are mutually contained in each others k-neighborhood.

source

CoherentStructures.neighborhood — Type.

neighborhood(ε)

Defines the ε-neighborhood sparsification method. In the final graph Laplacian, only those particle pairs are included which have distance less than ε.

source

Other sparsification methods can be implemented by defining a corresponding sparseaffinitykernel instance.

Diffusion-maps type graph Laplacian methods

CoherentStructures.diff_op — Function.

diff_op(data, sp_method, kernel = gaussian_kernel; α=1.0, metric=Euclidean"()")

Return a diffusion/Markov matrix P.

Arguments

data: 2D array with columns correspdonding to data points;

sp_method: employed sparsification method (neighborhood or mutualKNN);

kernel: diffusion kernel, e.g., x -> exp(-x*x/4σ);

α: exponent in diffusion-map normalization;

metric: distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$.

source

CoherentStructures.sparse_diff_op_family — Function.

sparse_diff_op_family(data, sp_method, kernel=gaussian_kernel, dim=2; op_reduce, α, metric)

Return a list of sparse diffusion/Markov matrices P.

Arguments

data: 2D array with columns correspdonding to data points;

sp_method: sparsification method;

kernel: diffusion kernel, e.g., x -> exp(-x*x/4σ);

dim: the columns are interpreted as concatenations of dim- dimensional points, to which metric is applied individually;

op_reduce: time-reduction of diffusion operators, e.g. mean or P -> prod(LMs.LinearMap,Iterators.reverse(P)) (default)

α: exponent in diffusion-map normalization;

metric: distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$.

source

CoherentStructures.sparse_diff_op — Function.

sparse_diff_op(data, sp_method, kernel; α=1.0, metric=Euclidean()) -> SparseMatrixCSC

Return a sparse diffusion/Markov matrix P.

Arguments

data: 2D array with columns correspdonding to data points;

sp_method: sparsification method;

kernel: diffusion kernel, e.g., x -> exp(-x*x) (default);

α: exponent in diffusion-map normalization;

metric: distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$.

source

CoherentStructures.sparseaffinitykernel — Function.

sparseaffinitykernel(data, sp_method, kernel, metric=Euclidean()) -> SparseMatrixCSC

Return a sparse matrix W where $w_{ij} = k(x_i, x_j)$. The $x_i$ are taken from the columns of data. Entries are only calculated for pairs determined by the sparsification method sp_method. Default metric is Euclidean().

source

CoherentStructures.kde_normalize! — Function.

kde_normalize!(A, α = 1.0)

Normalize rows and columns of A in-place with the respective row-sum to the α-th power; i.e., return $a_{ij}:=a_{ij}/q_i^{\alpha}/q_j^{\alpha}$, where $q_k = \sum_{\ell} a_{k\ell}$. Default for α is 1.0.

source

CoherentStructures.wLap_normalize! — Function.

wLap_normalize!(A)

Normalize rows of A in-place with the respective row-sum; i.e., return $a_{ij}:=a_{ij}/q_i$.

source

Adjancency-matrix-based graph Laplacian methods

CoherentStructures.sparse_adjacency — Function.

sparse_adjacency(data, ε[, dim]; metric) -> SparseMatrixCSC

Return a sparse adjacency matrix A with integer entries 0 or 1. If the third argument dim is passed, then data is interpreted as concatenated points of length dim, to which metric is applied individually. Otherwise, metric is applied to the whole columns of data.

Arguments

data: 2D array with columns correspdonding to data points;

ε: distance threshold;

dim: the columns of data are interpreted as concatenations of dim- dimensional points, to which metric is applied individually;

metric: distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$.

source

CoherentStructures.sparse_adjacency_list — Function.

sparse_adjacency_list(data, ε; metric=Euclidean()) -> idxs::Vector{Vector}

Return two lists of indices of data points that are adjacent.

Arguments

data: 2D array with columns correspdonding to data points;

ε: distance threshold;

metric: distance function w.r.t. which the kernel is computed, however, only for point pairs where $metric(x_i, x_j)\leq \varepsilon$.

source

Diffusion-coordinate-like functions

CoherentStructures.diffusion_coordinates — Function.

diffusion_coordinates(P,n_coords) -> (Σ::Vector, Ψ::Matrix)

Compute the (time-coupled) diffusion coordinates Ψ and the coordinate weights Σ for a linear map P. n_coords determines the number of diffusion coordinates to be computed.

source

CoherentStructures.diffusion_distance — Function.

diffusion_distance(diff_coord) -> SymmetricMatrix

Returns the distance matrix of pairs of points whose diffusion distances correspond to the diffusion coordinates given by diff_coord.

source