Geodesic vortices

Geodesic elliptic material vortices

The following functions implement an LCS methodology developed in the following papers:

Haller & Beron-Vera, 2012

Haller & Beron-Vera, 2013

Karrasch, Huhn, and Haller, 2015

The present code was originally inspired by Alireza Hadjighasem's MATLAB implementation, which was written in the context of the SIAM Review paper, but has been significantly modified and improved throughout. Depending on the indefinite metric tensor field used, the functions below yield the following types of coherent structures:

black-hole/Lagrangian coherent vortices (Haller & Beron-Vera, 2012)

elliptic objective Eulerian coherent structures (OECSs) (Serra & Haller, 2016)

material diffusive transport barriers (Haller, Karrasch, and Kogelbauer, 2018)

The general procedure is the following. Assume $T$ is the symmetric tensor field of interest, say, (i) the Cauchy-Green strain tensor field $C$, (ii) the rate-of-strain tensor field $S$, or (iii) the averaged diffusion-weighted Cauchy-Green tensor field $\bar{C}_D$; cf. the references above. Denote by $0<\lambda_1\leq\lambda_2$ the eigenvalue and by $\xi_1$ and $\xi_2$ the corresponding eigenvector fields of $T$. Then the direction fields of interest are given by

\[\eta_{\lambda}^{\pm} := \sqrt{\frac{\lambda_2 - \lambda}{\lambda_2-\lambda_1}}\xi_1 \pm \sqrt{\frac{\lambda - \lambda_1}{\lambda_2-\lambda_1}}\xi_2.\]

Tensor singularities are defined as points at which $\lambda_2=\lambda_1$, i.e., at which the two characteristic directions $\xi_1$ and $\xi_2$ are not well-defined. As described and exploited in Karrasch et al., 2015, non-negligible tensor singularities express themselves by an angle gap when tracking (the angle of) tensor eigenvector fields along closed paths surrounding the singularity. Our approach here avoids computing singularities directly, but rather computes the index for each grid cell and then combines nearby singularities, i.e., adds non-vanishing indices of nearby grid cells.

In summary, the implementation consists of the following steps:

compute the index for each grid cell and combine nearby singular grid cells to "singularity candidates";

look for elliptic singularity candidates (and potentially isolated wedge pairs);

place an eastwards oriented Poincaré section at the pair center;

for each point on the discretized Poincaré section, scan through the given parameter interval such that the corresponding η-orbit closes at that point;

if desired: for each Poincaré section, take the outermost closed orbit as the coherent vortex barrier (if there exist any).

Function documentation

The meta-functions ellipticLCS and constrainedLCS

The fully automated high-level functions are:

ellipticLCS
constrainedLCS

One of their arguments is a list of parameters used in the LCS detection. This list is combined in a data type called LCSParameters. The output is a list of EllipticBarriers and a list of Singularitys. There is an option to retrieve all closed barriers (outermost=false), in contrast to extracting only the outermost vortex boundaries (outermost=true), which is more efficient.

The meta-functions consist of two steps: first, the index theory-based determination of where to search for closed orbits,, cf. Index theory-based placement of Poincaré sections; second, the closed orbit computation, cf. Closed orbit detection.

Specific types

These are the specifically introduced types for elliptic LCS computations.

CoherentStructures.LCSParameters — Type.

struct LCSParameters

Container for parameters used in elliptic LCS computations.

Fields

indexradius::Float64=0.1: radius for singularity type detection

boxradius::Float64=0.5: "radius" of localization square for closed orbit detection

combine_pairs=true: whether isolated singularity pairs should be merged

n_seeds::Int64=40: number of seed points on the Poincaré section

pmin::Float64=0.7: lower bound on the parameter in the $\eta$-field

pmax::Float64=1.3: upper bound on the parameter in the $\eta$-field

rdist::Float64=1e-4: required return distances for closed orbits

source

CoherentStructures.EllipticBarrier — Type.

struct EllipticBarrier

This is a container for coherent vortex boundaries. An object vortex of type EllipticBarrier can be easily plotted by plot(vortex.curve), or plot!([figure, ]vortex.curve) if it is to be overlaid over an existing plot.

Fields

curve: a list of tuples, contains the coordinates of coherent vortex boundary points;

core: location of the vortex core;

p: contains the parameter value of the direction field $\eta_{\lambda}^{\pm}$, for the curve is a closed orbit;

s: a Bool value, which encodes the sign in the formula of the direction field $\eta_{\lambda}^{\pm}$ via the formula $(-1)^s$.

source

Another one is Singularity, which comes along with some convenience functions.

Singularity
getcoords
getindices

Index theory-based placement of Poincaré sections

This is performed by discrete_singularity_detection for line fields (such as eigenvector fields of symmetric positive-definit tensor fields) and by critical_point_detection for classic vector fields.

CoherentStructures.singularity_detection — Function.

singularity_detection(T, combine_distance; combine_isolated_wedges=true) -> Vector{Singularity}

Calculates line-field singularities of the first eigenvector of T by taking a discrete differential-geometric approach. Singularities are calculated on each cell. Singularities with distance less or equal to combine_distance are combined by averaging the coordinates and adding the respective indices. If combine_pairs is `true, pairs of singularities that are mutually the closest ones are included in the final list.

Returns a vector of Singularitys. Returned indices correspond to doubled indices to get integer values.

source

CoherentStructures.critical_point_detection — Function.

critical_point_detection(vs, combine_distance, γ; combine_pairs=true)

Computes critical points of a vector/line field vs, given as an AxisArray. Critical points with distance less or equal to combine_distance are combined by averaging the coordinates and adding the respective indices. The parameter γ should be chosen π for line fields and 2π for vector fields; cf. compute_singularities. If combine_pairs istrue, pairs of singularities that are mutually the closest ones are included in the final list.

source

This function takes three steps.

CoherentStructures.compute_singularities — Function.

compute_singularities(α, modulus) -> Vector{Singularity}

Computes critical points/singularities of vector and line fields, respectively. α is a scalar field (array) which is assumed to contain some consistent angle representation of the vector/line field. Choose modulus=2π for vector fields, and modulus=π for line fields.

source

CoherentStructures.combine_singularities — Function.

combine_singularities(sing_coordinates, sing_indices, combine_distance) -> Vector{Singularity}

This function does the equivalent of: Build a graph where singularities are vertices, and two vertices share an edge iff the coordinates of the corresponding vertices (given by sing_coordinates) have a distance leq combine_distance. Find all connected components of this graph, and return a list of their mean coordinate and sum of sing_indices

source

CoherentStructures.combine_isolated_pairs — Function.

combine_isolated_pairs(singularities)

Determines singularities which are mutually closest neighbors and combines them as one, while adding their indices.

source

From all virtual/merged singularities those with a suitable index are selected. Around each elliptic singularity the tensor field is localized and passed on for closed orbit detection.

Closed orbit detection

CoherentStructures.compute_returning_orbit — Function.

compute_returning_orbit(vf, seed::SVector{2}, save::Bool=false)

Computes returning orbits under the velocity field vf, originating from seed. The optional argument save controls whether intermediate locations of the returning orbit should be saved.

source

CoherentStructures.compute_closed_orbits — Function.

compute_closed_orbits(ps, ηfield, cache; rev=true, pmin=0.7, pmax=1.5, rdist=1e-4)

Compute the outermost closed orbit for a given Poincaré section ps, a vector field constructor ηfield, and an LCScache cache. Keyword argumentspminandpmaxcorrespond to the range of shift parameters in which closed orbits are sought;revdetermines whether closed orbits are sought from the outside inwards (true) or from the inside outwards (false).rdist` sets the required return distance for an orbit to be considered as closed.

source