FEM-based methods

FEM-based Methods

These methods rely on the theory outlined by Froyland's Dynamical Laplacian and the Geometric Heat Flow of Karrasch & Keller.

The Laplace-like operators are best discretized by finite-element-based methods, see this paper by Froyland & Junge.

This involves the discretization of the average of a one-parameter family of Laplace operators of the form:

\[\Delta^{dyn} := \sum_{t \in \mathcal T} P_t^* \Delta P_t\]

for a finite series of times $\mathcal T$, where $P_t$ is the transfer-operator for the flow at time $t$ (in volume-preserving flows).

The resulting operator is both symmetric and uniformly elliptic. Eigenfunctions of $\Delta^{dyn}$ can be used to find Lagrangian Coherent Structures.

See the Examples section for examples of how these methods can be used.

Features

CG and TO methods

The standard Galerkin formulation of the weak dynamical Laplace is refered to as the CG-method here, due to the fact that the inverse Cauchy-Green tensor appears in the weak formulation. This gives a bilinear form $\overline a(u,v) := \sum_{t \in \mathcal T}a^t(P_t u, P_t v)$ Here $P_t$ is the Transfer-Operator (or pushforward) to time-$t$, and $a^t$ is the weak-form of the Laplacian on the range of the time-$t$ map being considered. There are also a range of Transfer-Operator based approaches implemented here. These approximate the weak form of the Dynamical-Laplace by a bilinear-form:

\[\tilde a_h(u,v) = \sum_{t \in \mathcal T} a^t(I_hP_t u, I_h P_t v)\]

where $I_h$ is a suitable interpolation operator depending on the mesh-width $h$. Options for $I_h$ implemented in this package are:

Collocation (pointwise interpolation)...

Points used are mesh points from domain grid ("adaptive TO")

Points usedare arbitrary("non-adaptive TO")

the $L^2$-orthogonal projection onto a FEM-space

Using the forwards flow map (currently gives poor results)

Using the inverse flow map

Note that the $L^2$-Galerkin methods currently perform very poorly on larger problems.

For more details, see this paper.

Grids

Note that irregular meshes currently do not work in Julia 1.0, the documentation has not been changed to reflect this.

Various types of regular and irregular meshes (with Delaunay triangulation using VoronoiDelaunay.jl ) are supported. These are based on the corresponding elements from JuAFEM.jl and include:

Triangular P1-Lagrange elements in 2D (all methods)

Quadrilateral P1-Lagrange elements in 2D (all methods except adaptive TO)

Triangular and Quadrilateral P2-Lagrange elements in 2D (all methods except adaptive TO)

Tetrahedral P1-Lagrange elements in 3D (only CG method tested, non-adaptive TO might work also)

The gridContext Type

The FEM-based methods of CoherentStructures.jl rely heavily on the JuAFEM.jl package. This package is very low-level and does not provide point-location/plotting functionality. To be able to more conveniently work with the specific types of grids that we need, all necessary variables for a single grid are combined in a gridContext structure - including the grid points, the quadrature formula used and the type of element used (e.g. Triangular P1, Quadrilateral P2, etc..). This makes it easier to assemble mass/stiffness matrices, and provides an interface for point-location and plotting.

In this documentation, the variable name ctx is exclusively used for gridContext objects.

See also Constructing Grids in the FEM-API section.

Node ordering and dof ordering

Finite Element methods work with degrees of freedom (dof), which are elements of some dual space. For nodal finite elements, these correspond to evaluation functionals at the nodes of the grid.

The nodes of the grid can be obtained in the following way [n.x for n in ctx.grid.nodes]. However, most of the methods of this package do not return results in this order, but instead use JuAFEM.jl's dof-ordering.

See also the documentation in dof2node and CoherentStructures.gridContext

When working with (non-natural) Boundary Conditions, the ordering is further changed, due to there being fewer degrees of freedom in total.

Assembly

See Stiffness and Mass Matrices from the FEM-API section.

Evaluating Functions in the Approximation Space

given a series of coefficients that represent a function in the approximation space, to evaluate a function at a point, use the evaluate_function_from_nodevals or evaluate_function_from_dofvals functions.

using Plots
ctx = regularP2TriangularGrid((10,10))
u = zeros(ctx.n)
u[45] = 1.0
Plots.heatmap(range(0,stop=1,length=200),range(0,stop=1,length=200), (x,y)->evaluate_function_from_nodevals(ctx,u,[x,y]))

For more details, consult the API: evaluate_function_from_dofvals, evaluate_function_from_nodevals

Nodal Interpolation

To perform nodal interpolation of a grid, use the nodal_interpolation function.

Boundary Conditions

To use something other than the natural homogeneous von Neumann boundary conditions, the CoherentStructures.boundaryData type can be used. This currently supports combinations of homogeneous Dirichlet and periodic boundary conditions.

Homogeneous Dirichlet BCs require rows and columns of the stiffness/mass matrices to be deleted

Periodic boundary conditions require rows and columns of the stiffness/mass matrices to be added to each other.

This means that the coefficient vectors for elements of the approximation space that satisfy the boundary conditions are potentially smaller and in a different order. Given a bdata argument, functions like plot_u will take this into account.

Constructing Boundary Conditions

Natural von-Neumann boundary conditions can be constructed with: boundaryData() and are generally the default

Homogeneous Dirichlet boundary conditions can be constructed with the getHomDBCS(ctx,[which="all"]) function. The optional which parameter is a vector of strings, corresponding to JuAFEM face-sets, e.g. getHomDBCS(ctx,which=["left","right"])

Periodic boundary conditions are constructed by calling boundaryData(ctx,predicate,[which_dbc=[]]). The argument predicate is a function that should return true if and only if two points should be identified. Due to floating-point rounding errors, note that using exact comparisons (==) should be avoided. Only points that are in JuAFEM.jl boundary facesets are considered. If this is too restrictive, use the boundaryData(dbc_dofs, periodic_dofs_from,periodic_dofs_to) constructor.

For details, see boundaryData

Example

Here we apply Homogeneous DBC to top and bottom, and identify the left and right side:

using CoherentStructures
ctx = regularQuadrilateralGrid((10,10))
predicate = (p1,p2) -> abs(p1[2] - p2[2]) < 1e-10 && peuclidean(p1[1],p2[1],1.0) < 1e-10
bdata = boundaryData(ctx,predicate,["top","bottom"])
u = ones(nDofs(ctx,bdata))
u[20] = 2.0; u[38] = 3.0; u[56] = 4.0
plot_u(ctx,u,200,200,bdata=bdata,colorbar=:none)

To apply boundary conditions to a stiffness/mass matrix, use the applyBCS function. Note that assembleStiffnessMatrix and assembleMassMatrix take a bdata argument that does this internally.

Plotting and Videos

There are some helper functions that exist for making plots and videos of functions on grids. These rely on the Plots.jl library. Plotting recipes are unfortunately not implemented.

The simplest way to plot is using the plot_u function. Plots and videos of eulerian plots like $f \circ \Phi^0_t$ can be made with the plot_u_eulerian and eulerian_videos functions.

Parallelisation

Many of the plotting functions support parallelism internally. Tensor fields can be constructed in parallel, and then passed to assembleStiffnessMatrix. For an example that does this, see TODO: Add this example

FEM-API

Stiffness and Mass Matrices

CoherentStructures.assembleStiffnessMatrix — Function.

assembleStiffnessMatrix(ctx,A,[p; bdata])

Assemble the stiffness-matrix for a symmetric bilinear form

\[a(u,v) = \int \nabla u(x)\cdot A(x)\nabla v(x)f(x) dx\]

The integral is approximated using quadrature. A is a function that returns a Tensors.SymmetricTensor{2,dim} and has one of the following forms:

A(x::Vector{Float64})

A(x::Vec{dim})

A(x::Vec{dim}, index::Int, p). Here x is equal to ctx.quadrature_points[index], and p is that which is passed to assembleStiffnessMatrix

The ordering of the result is in dof order, except that boundary conditions from bdata are applied. The default is natural boundary conditions.

source

CoherentStructures.assembleMassMatrix — Function.

assembleMassMatrix(ctx;[bdata,lumped=false])

Assemble the mass matrix

\[M_{i,j} = \int \varphi_j(x) \varphi_i(x) f(x)d\lambda^d\]

The integral is approximated using numerical quadrature. The values of f(x) are taken from ctx.mass_weights, and should be ordered in the same way as ctx.quadrature_points

The result is ordered in a way so as to be usable with a stiffness matrix with boundary data bdata.

Returns a lumped mass matrix if lumped==true.

Example

ctx.mass_weights = map(f, ctx.quadrature_points)
M = assembleMassMatrix(ctx)

source

Constructing Grids

There are several helper functions available for constructing grids. The simplest is:

CoherentStructures.regular2DGrid — Function.

regular2DGrid(gridType, numnodes, LL=[0.0,0.0],UR=[1.0,1.0];quadrature_order=default_quadrature_order)

Constructs a regular grid. gridType should be from CoherentStructures.regular2DGridTypes

source

Supported values for the gridType argument are:

CoherentStructures.regular2DGridTypes

6-element Array{String,1}:
 "regular triangular grid"
 "regular P2 triangular grid"
 "regular Delaunay grid"
 "regular P2 Delaunay grid"
 "regular quadrilateral grid"
 "regular P2 quadrilateral grid"

The following functions are conceptually similar:

CoherentStructures.regularTriangularGrid — Function.

regularTriangularGrid(numnodes=(25,25), LL=[0.0,0.0],UR=[1.0,1.0], quadrature_order=default_quadrature_order)

Create a regular P1 triangular grid on a rectangle; it does not use Delaunay triangulation internally.

source

CoherentStructures.regularP2TriangularGrid — Function.

regularP2TriangularGrid(numnodes=(25,25), LL=[0.0,0.0],UR=[1.0,1.0],quadrature_order=default_quadrature_order)

Create a regular P2 triangular grid on a Rectangle. Does not use Delaunay triangulation internally.

source

CoherentStructures.regularQuadrilateralGrid — Function.

regularP2QuadrilateralGrid(numnodes=(25,25), LL=[0.0,0.0],UR=[1.0,1.0],quadrature_order=default_quadrature_order)

Create a regular P1 quadrilateral grid on a Rectangle.

source

CoherentStructures.regularP2QuadrilateralGrid — Function.

regularP2QuadrilateralGrid(numnodes=(25,25), LL=[0.0,0.0], UR=[1.0,1.0], quadrature_order=default_quadrature_order)

Create a regular P2 quadrilateral grid on a rectangle.

source

In 3D we have

CoherentStructures.regularTetrahedralGrid — Function.

regularTetrahedralGrid(numnodes=(10,10,10), LL=[0.0,0.0,0.0], UR=[1.0,1.0,1.0], quadrature_order=default_quadrature_order3D)

Create a regular P1 tetrahedral grid on a Cuboid in 3D. Does not use Delaunay triangulation internally.

source

CoherentStructures.regularP2TetrahedralGrid — Function.

regularP2TetrahedralGrid(numnodes=(10,10,10), LL=[0.0,0.0,0.0], UR=[1.0,1.0,1.0], quadrature_order=default_quadrature_order3D)

Create a regular P2 tetrahedral grid on a Cuboid in 3D. Does not use Delaunay triangulation internally.

source

All of these methods return a gridContext object.

CoherentStructures.gridContext — Type.

struct gridContext<dim>

Stores everything needed as "context" to be able to work on a FEM grid based on the JuAFEM package. Adds a point-locator API which facilitates plotting functions defined on the grid within Julia.

Fields

grid::JuAFEM.Grid, ip::JuAFEM.Interpolation, qr::JuAFEM.QuadratureRule - See the JuAFEM package

loc::CellLocator object used for point-location on the grid.

node_to_dof::Vector{Int} lookup table for dof index of a node

dof_to_node::Vector{Int} inverse of nodetodof

n::Int number of nodes on the grid

m::Int number of elements (e.g. triangles,quadrilaterals, ...) on the grid

quadrature_points::Vector{Vec{dim,Float64}} All quadrature points on the grid, in a fixed order.

mass_weights::Vector{Float64} Weighting for mass matrix

spatialBounds If available, the corners of a bounding box of a domain. For regular grids, the bounds are tight.

numberOfPointsInEachDirection For regular grids, how many (non-interior) nodes make up the regular grid.

gridType A string describing what kind of grid this is (e.g. "regular triangular grid")

source

Irregular grids

The constructors for CoherentStructures.gridContext, including one for irregular Delaunay grids, are not exported by default, the documentation is available through the REPL:

help?> (::Type{CoherentStructures.gridContext{2}})

Boundary Conditions API

CoherentStructures.boundaryData — Type.

mutable struct boundaryData

Represent (a combination of) homogeneous Dirichlet and periodic boundary conditions. Fields:

dbc_dofs list of dofs that should have homogeneous Dirichlet boundary conditions. Must be sorted.

periodic_dofs_from and periodic_dofs_to are both Vector{Int}. The former must be strictly increasing, both must be the same length. periodic_dofs_from[i] is identified with periodic_dofs_to[i]. periodic_dofs_from[i] must be strictly larger than periodic_dofs_to[i]. Multiple dofs can be identified with the same dof. If some dof is identified with another dof and one of them is in dbc_dofs, both points must be in dbc_dofs

source

CoherentStructures.getHomDBCS — Function.

getHomDBCS(ctx,which="all")

Return boundaryData object corresponding to homogeneous Dirichlet Boundary Conditions for a set of facesets. which="all" is shorthand for ["left","right","top","bottom"].

source

CoherentStructures.undoBCS — Function.

undoBCS(ctx,u,bdata)

Given a vector u in dof order with boundary conditions applied, return the corresponding u in dof order without the boundary conditions.

source

CoherentStructures.applyBCS — Function.

applyBCS(ctx,K,bdata)

Apply the boundary conditions from bdata to the ctx.n by ctx.n sparse matrix K.

source

Helper functions

CoherentStructures.dof2node — Function.

dof2node(ctx,u)

Interprets u as an array of coefficients ordered in dof order, and reorders them to be in node order.

source

CoherentStructures.getDofCoordinates — Function.

getDofCoordinates(ctx,dofindex)

Return the coordinates of the node corresponding to the dof with index dofindex

source

CoherentStructures.evaluate_function_from_dofvals — Function.

evaluate_function_from_dofvals(ctx,dofvals,x_in; [outside_value=0,project_in=false])

Evaluate a function in the approximation space at the point x_in. If x_in is out of points, return outside_value. If project_in is true, points not within ctx.spatialBounds are first projected into the domain.

The coefficients in nodevals are interpreted to be in dof order.

source

CoherentStructures.evaluate_function_from_nodevals — Function.

evaluate_function_from_nodevals(ctx,nodevals,x_in; [outside_value=0, project_in=false])

Like evaluate_function_from_dofvals, but the coefficients from nodevals are assumed to be in node order.

source

CoherentStructures.nodal_interpolation — Function.

nodal_interpolation(ctx,f)

Perform nodal interpolation of a function. Returns a vector of coefficients in dof order

source

CoherentStructures.getH — Function.

getH(ctx)

Return the mesh width of a regular grid.

source

Plotting API

FEM

CoherentStructures.plot_u — Function.

plot_u(ctx, dof_vals, nx, ny; bdata=nothing, kwargs...)

Plot the function with coefficients (in dof order, possible boundary conditions in bdata) given by dof_vals on the grid ctx. The domain to be plotted on is given by ctx.spatialBounds. The function is evaluated on a regular nx by ny grid, the resulting plot is a heatmap. Keyword arguments are passed down to plot_u_eulerian, which this function calls internally.

source

CoherentStructures.plot_u_eulerian — Function.

plot_u_eulerian(ctx,dof_vals,inverse_flow_map,
    LL,UR,nx,ny,
    euler_to_lagrange_points=nothing, only_get_lagrange_points=false,
    z=nothing,
    postprocessor=nothing,
    bdata=nothing, ....)

Plot a heatmap of a function in eulerian coordinates, i.e. the pushforward of $f$. This is given by $f \circ \Phi^{-1}$, $f$ is a function defined on the grid ctx, represented by coefficients given by dof_vals (with possible boundary conditions given in bdata)

The argument inverse_flow_map is $\Phi^{-1}$.

The resulting plot is on a regular nx by ny grid on the grid with lower left corner LL and upper right corner UR.

Points that fall outside of the domain represented by ctx are plotted as NaN, which results in transparency.

One can pass values to be plotted directly by providing them in an array in the argument z. postprocessor can modify the values being plotted, return_scalar_field results in these values being returned. See the source code for further details. Additional arguments are passed to Plots.heatmap

Inverse flow maps are computed in parallel if there are multiple workers.

source

CoherentStructures.eulerian_videos — Function.

 eulerian_videos(ctx, us, inverse_flow_map_t, t0,tf, nx,ny,nt, LL,UR, num_videos=1;
    extra_kwargs_fun=nothing, ...)

Create num_videos::Int videos in eulerian coordinates, i.e. where the time $t$ is varied, plot $f_i \circ \Phi_t^0$ for $f_1, \dots$.

us(i,t) is a vector of dofs to be plotted at time t for the ith video.

inverse_flow_map_t(t,x) is $\Phi_t^0(x)$

t0, tf are initial and final time. Spatial bounds are given by LL,UR

nx,ny,nt give the number of points in each direction.

extra_kwargs_fun(i,t) can be used to provide additional keyword arguments to Plots.heatmap()

Additional kwargs are passed down to plot_eulerian_video

As much as possible is done in parallel.

Returns a Vector of iterables result. Call Plots.animate(result[i]) to get an animation.

source

CoherentStructures.eulerian_video — Function.

eulerian_video(ctx, u, inverse_flow_map_t,t0,tf, nx, ny, nt, LL, UR;extra_kwargs_fun=nothing,...)

Like eulerian_videos, but u(t) is a vector of dofs, and extra_kwargs_fun(t) gives extra keyword arguments. Returns only one result, on which Plots.animate() can be applied.

source

Other plotting utilities

CoherentStructures.plot_ftle — Function.

plot_ftle(odefun,p,tspan,LL,UR,nx,ny;
    δ=1e-9,tolerance=1e-4,solver=OrdinaryDiffEq.BS5(),
    existing_plot=nothing,flip_y=false, check_inbounds=always_true)

Make a heatmap of a FTLE field using finite differences. If existing_plot is given a value, plot using heatmap! on top of it. If flip_y is true, then flip the y-coordinate (needed sometimes due to a bug in Plots). Points where check_inbounds(x[1],x[2],p) == false are set to NaN (i.e. transparent).

source

Defaults

const default_quadrature_order=5
const default_solver = OrdinaryDiffEq.BS5()